% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/illustrate_powers.R
\name{illustratePowerMultipleTests}
\alias{illustratePowerMultipleTests}
\title{Runs simulations of sample size and delta to investigate the power of multiple equivalence tests run simultaneously}
\usage{
illustratePowerMultipleTests(pi, alpha = 0.05, ns = round(10^(seq(1.67,
  3.33, by = 0.33))), deltas = c(0.01, 0.02, 0.03, 0.04, 0.05, 0.075, 0.1),
  Nsim = 100, plot = TRUE, verbose = FALSE)
}
\arguments{
\item{pi}{A list of length L where the entries are the true, gold-standard probability 
vectors which define the population characteristics. We are simulating under 
these scenarios to see how often we can detect them while preserving the
family-wise error rate (i.e., the test power).}

\item{alpha}{The minimum family-wise Type I error rate (test size). Default is \code{5\%}.}

\item{ns}{Which sample sizes should be simulated? The default is \code{round(10^(seq(1.67, 3.33, by = 0.33)))}
corresponding to \code{47, 100, 214, 457, 977, 2089}.}

\item{deltas}{Which deltas (the practical equivalence range) should be simulated? The default
are the scalars \code{0.01, 0.02, 0.03, 0.04, 0.05, 0.075, 0.10}.}

\item{Nsim}{How many simulations per n and delta combination? The higher the value, the
greater resolution. Default is \code{100} for speed.}

\item{plot}{Should we plot power by n and delta upon completion? Default is \code{TRUE}.}

\item{verbose}{Should we print out messages indicating progress during a potentionally
long simulation? Default is \code{FALSE}.}
}
\value{
Simulated powers in a matrix of dimension number of 
							\code{ns} x number of \code{deltas}.
}
\description{
Runs simulations of sample size and delta to investigate the power of multiple equivalence tests run simultaneously
}
\examples{
 \dontrun{
	pi_gender = c(0.5, 0.5)
	pi_ethnicity = c(.13, .70, .13, .04)
	pi_age = c(.13, .41, .30, .16)
	pi_region = c(.22, .18, .37, .23)
	illustratePowerMultipleTests(
			pi = list(pi_gender, pi_ethnicity, pi_age, pi_region), 
			Nsim = 500, verbose = TRUE)
	}
}
\author{
Adam Kapelner
}

