% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_representativeness_tests.R
\name{runMultipleRepresentativenessTests}
\alias{runMultipleRepresentativenessTests}
\title{Runs a set of L Equivalence Hypothesis Tests with the null of non-representativeness for a fixed sample size.
This makes use of the \code{runSingleRepresentativenessTest} function.}
\usage{
runMultipleRepresentativenessTests(pi, x, delta, alpha = 0.05,
  illustrate = TRUE)
}
\arguments{
\item{pi}{A list of length L where the entries are the true, gold-standard probability 
vectors which define the population characteristics.
We are testing against these to ascertain if the sample is representative.}

\item{x}{A list of length L where the entries are the experimental samples consisting of a count 
vector. The length of each vector must be equal to the corresponding vector length in \code{pi}.}

\item{delta}{The practical equivalence tolerance level. This can be a scalar (implying
a d-length vector with the same value) or a vector of length d.}

\item{alpha}{The minimum family-wise Type I error rate (test size). Default is \code{5\%}.}

\item{illustrate}{Should each of the L tests be illustrated? For each test, each of the d cells can be 
illustrated as a bar plot showing pi +- delta with the p-hat sample and its 
confidence interval. Default is code{TRUE}.}
}
\value{
A list which contains function arguments as well as the following keys: 
							(a) reject_null - a boolean which indicates
							if the non-representativeness null was rejected
							(b) ci - this is the simultaneous confidence interval for the
							multinomial sample proportion vector as a d x 2 matrix 
							(c) pval - the significance level for the family of tests
							(d) pvals - the significance levels for each test using the 
							Bonferroni-Holm correction
}
\description{
Runs a set of L Equivalence Hypothesis Tests with the null of non-representativeness for a fixed sample size.
This makes use of the \code{runSingleRepresentativenessTest} function.
}
\examples{
 \dontrun{
	pi_gender = c(0.5, 0.5)
	pi_ethnicity = c(.13, .70, .13, .04)
	pi_age = c(.13, .41, .30, .16)
	pi_region = c(.22, .18, .37, .23)
	#pretend to collect data
	n = 1000
	x_gender = rmultinom(1, n, pi_gender)
	x_ethnicity = rmultinom(1, n, pi_ethnicity)
	x_age = rmultinom(1, n, pi_age)
	x_region = rmultinom(1, n, pi_region)
	res = runMultipleRepresentativenessTests(
				list(pi_gender, pi_ethnicity, pi_age, pi_region),
				list(x_gender, x_ethnicity, x_age, x_region),  
				delta = 0.03,
				alpha = 0.05)
	res$pval
	#you can also see this by running all tests and then doing 
	#the p-val correction at the end
	res1 = runSingleRepresentativenessTest(pi_gender, x_gender, 
				delta = 0.03, alpha = 0.05)
	res2 = runSingleRepresentativenessTest(pi_ethnicity, x_ethnicity, 
				delta = 0.03, alpha = 0.05)
	res3 = runSingleRepresentativenessTest(pi_age, x_age, 
				delta = 0.03, alpha = 0.05)
	res4 = runSingleRepresentativenessTest(pi_region, x_region, 
				delta = 0.03, alpha = 0.05)
	p.adjust(p = c(res1$pval, res2$pval, res3$pval, res4$pval), method = "holm")
	}
}
\author{
Adam Kapelner
}

