% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleRepresentativenessChecker.R
\docType{package}
\name{SampleRepresentativenessChecker}
\alias{SampleRepresentativenessChecker}
\alias{SampleRepresentativenessChecker-package}
\title{Sample Representativeness Checker}
\description{
Uses an equivalence hypothesis test to to check representativeness
of a sample for a population of interest given known cell probabilities in the
population. Includes confidence interval illustration and power illustrations
assuming a practical equivalence level.
}
\author{
Adam Kapelner \email{kapelner@qc.cuny.edu}
}
\references{
TK
}
\keyword{Sample}
\keyword{representativeness}
\keyword{survey,}

